+++ src/adapter/flow-dataset/src/flow_controllers/flow_controller_transform.rs
@@ -126,11 +126,11 @@ impl fs::FlowController for FlowControllerTransform {
                 let activation_cause = fs::FlowActivationCause::ResourceUpdate(
                     fs::FlowActivationCauseResourceUpdate {
                         activation_time: finish_time,
-                        changes: fs::ResourceChanges::NewData {
+                        changes: fs::ResourceChanges::NewData(fs::ResourceDataChanges {
                             blocks_added: dataset_increment.num_blocks,
                             records_added: dataset_increment.num_records,
                             new_watermark: dataset_increment.updated_watermark,
-                        },
+                        }),
                         resource_type: DATASET_RESOURCE_TYPE.to_string(),
                         details: serde_json::to_value(DatasetResourceUpdateDetails {
                             dataset_id,