+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Recommendation: for ease of reading, use the following order:
 ## Unreleased
 ### Changed
 - dep: `ringbug` updated from `0.3` to `0.4`.
+- Flows: each input contribution to batching rule is reflected as an update of start condition,
+   so that flow history may show how many accumulated records were present at the moment of each update
 
 ## [0.247.0] - 2025-08-28
 ### Added